updated and fixed hooks, styles, modal fixes
