+++ client/styles/forging.scss
@@ -2,7 +2,7 @@ $primary-color: #FF7A00;
 $background-color-modal: rgba(0, 0, 0, 0.5);
 $border-color-dropdown: rgba(255, 123, 0, 0.24);
 
-.TBG-page {
+.forging-page {
   display: flex;
   flex-direction: column;
   align-items: center;
