--- client/utils/modalContext.js
@@ -1,26 +1,0 @@
-import React, { createContext, useContext, useState } from 'react';
-
-const ModalContext = createContext();
-
-export const useModal = () => useContext(ModalContext);
-
-export const ModalProvider = ({ children }) => {
-  const [modalContent, setModalContent] = useState(null);
-  const [isOpen, setIsOpen] = useState(false);
-
-  const openModal = (content) => {
-    setModalContent(content);
-    setIsOpen(true);
-  };
-
-  const closeModal = () => {
-    setIsOpen(false);
-    setModalContent(null);
-  };
-
-  return (
-    <ModalContext.Provider value={{ isOpen, modalContent, openModal, closeModal }}>
-      {children}
-    </ModalContext.Provider>
-  );
-};
